requirements.txt
flask
sentence-transformers
faiss-cpu
requests
numpy
langchain-text-splitters
pyngrok
gunicorn
